cv2 expect float32

+++ python/src/kernelphysiology/dl/experiments/colour_discrimination/datasets/colour_spaces.py
@@ -81,7 +81,7 @@ def dkl012rgb01(x):
 
 def rgb2hsv01(x):
     x = _rgb2double(x)
-    x = np.float32(cv2.cvtColor(x, cv2.COLOR_RGB2HSV))
+    x = np.float32(cv2.cvtColor(x.astype('float32'), cv2.COLOR_RGB2HSV))
     x[:, :, 0] /= 360
     return x
 
@@ -92,5 +92,5 @@ def hsv012rgb(x):
 
 def hsv012rgb01(x):
     x[:, :, 0] *= 360
-    x = cv2.cvtColor(x, cv2.COLOR_HSV2RGB)
+    x = cv2.cvtColor(x.astype('float32'), cv2.COLOR_HSV2RGB)
     return _clip01(x)
